pyproject.toml
[project]
name = "mailchimp-image-processor"
version = "0.1.0"
description = "An application to upload images to Mailchimp's Content Studio."
readme = "README.md"
authors = [
    { name = "Ali Samji", email = "aisamji@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.1.0",
    "pyinstaller>=6.17.0",
]

[project.scripts]
mailchimp-image-processor = "mailchimp_image_processor:main"
mip = "mailchimp_image_processor:main"

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

